% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cm.r
\name{best_cm}
\alias{best_cm}
\title{Confusion Matrix for the Best Model}
\usage{
best_cm(CV = cv, auc_out = auc_out)
}
\arguments{
\item{CV}{Output object from loocv().}

\item{auc_out}{Output object form Cal_AUC()}
}
\value{
A list of confusionMatrix class.
}
\description{
Given an \code{loocv()} output and \code{Cal_AUC()} output, this function calculates the confusion matrix for the label produced from the model with optimized number of features.
}
\examples{
data(training)

### using leave-one-out cross-validation ###
cv = loocv(training)

auc_out = Cal_AUC(cv)

best_cm(CV=cv,auc_out=auc_out)

}
