% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_AUC.r
\name{Cal_AUC}
\alias{Cal_AUC}
\title{Calculate Area Under the ROC Curve from Leave-One-Out Cross-Validation}
\usage{
Cal_AUC(CV = cv)
}
\arguments{
\item{CV}{The output object from loocv()}
}
\value{
A data frame of 6 columns, comparison group1 label, comparison group2 label, number of features included in the model, AUC, AUC * prior probability, and selected features.
}
\description{
This function calculates the AUC for each selected subset of features from DMBC cross-validation results, and output the set of features which provide the greatest AUC.
}
\examples{
data(training)
Cal_AUC(loocv(training))
}
