% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalPrb.r
\name{CalPrb}
\alias{CalPrb}
\title{Calculate Likelihood based on Dirichlet-multinomial distribution estimated parameters}
\usage{
CalPrb(FS_out = FS_out, testSet = testSet, col_start = 3,
  type_col = 2, HighestRank = nrow(FS_out$Feature))
}
\arguments{
\item{FS_out}{An object from the FeatureSelection()}

\item{testSet}{A test set in data frame or matrix form. The colnames should have the same bacteria (features) as in the training set.}

\item{col_start}{An index indicating at which column is the beginning of bacteria (features) data. The default is the 3rd column.}

\item{type_col}{An index indicating at which column is group/type variable. The default is the 2nd column.}

\item{HighestRank}{The top number of features inclueded in model. The default is all the features left after filtering.}
}
\value{
A data frame with 17 columns, each row represents a model estimation output.
}
\description{
This function estimates parameters from Dirichlet-multinomial distribution.
}
\examples{
data(training)

#### Take one row as testSet ####
idx <- sample(1:nrow(training),1)
test <- training[idx,]
train <- training[-idx,]

CalPrb(FS(train),test) # This may take up to one minute
}
