% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{clusters,dmbc_config-method}
\alias{clusters,dmbc_config-method}
\alias{dmbc_config-clusters}
\title{Extract the final cluster memberships from a \code{dmbc_config} class instance.}
\usage{
\S4method{clusters}{dmbc_config}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_config}}.}

\item{newdata}{An object of no explicit specification (currently ignored).}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Extract the final cluster memberships from a \code{dmbc_config} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
