\name{dmbc-internal}
\alias{dmbc.init}
\alias{dmbc.logLik.rbmds}
\alias{dmbc.logLik}
\alias{log_marg_lik}
\alias{log_marg_prior}
\alias{log_corr_fact}
\alias{logit}
\alias{expit}
\alias{list2matrix}
\alias{list2array}
\alias{list.sum}
\alias{relabel.theta}
\alias{pack.par}
\alias{permutations}
\title{Internal Objects}
\description{
   Internal \code{dmbc} functions.
}
%\usage{
%}
%\details{
%  Most of these are not to be called by the user.
%}
%\arguments{
% \item{\dots}{Type help(package = "dmbc") to get started.}
%}
\seealso{
   \code{\link{dmbc-class}},
   \code{\link{dmbcIC-class}},
   \code{\link{dmbc}},
   \code{\link{dmbc.IC}}.
}
\keyword{internal}
