% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\docType{class}
\name{dmbc_fit_list-class}
\alias{dmbc_fit_list-class}
\alias{dmbc_fit_list}
\title{An S4 class to represent the results of fitting DMBC model.}
\description{
An S4 class to represent the results of fitting DMBC model using multiple
  Markov Chain Monte Carlo chains.
}
\section{Slots}{

\describe{
\item{\code{results}}{An object of class \code{list}; list of \code{dmbc_fit}
objects corresponding to the parallel MCMC chains simulated during the
estimation.}
}}

\examples{
showClass("dmbc_fit_list")

}
\references{
Venturini, S., Piccarreta, R. (2019), "A Bayesian Approach for Model-Based
  Clustering of Several Binary Dissimilarity Matrices: the \pkg{dmbc}
  Package in \code{R}", Technical report.
}
\seealso{
\code{\link{dmbc_fit}} for more details on the components of each element of
  the list.
}
