% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmbc_mcmc}
\alias{dmbc_mcmc}
\alias{dmbc-internal}
\alias{dmbc_internal}
\alias{dmbc_relabel}
\alias{dmbc_pack_par}
\title{Internal functions for MCMC simulation.}
\usage{
dmbc_mcmc(
  raiD,
  raix,
  raing,
  radalpha,
  rn,
  rp,
  rG,
  rS,
  rtotiter,
  radZ,
  rgamma_z,
  reta,
  rgamma_alpha,
  rsigma2,
  rlambda,
  rhyper_eta_a,
  rhyper_eta_b,
  rhyper_sigma2_a,
  rhyper_sigma2_b,
  rhyper_lambda,
  rfamily,
  rverbose
)

dmbc_relabel(
  radtheta,
  radz,
  radalpha,
  radeta,
  radsigma2,
  radlambda,
  radprob,
  raix_ind,
  rinit,
  rn,
  rp,
  rS,
  rM,
  rR,
  rG,
  rverbose
)

dmbc_pack_par(radz, radalpha, radlambda, rn, rp, rM, rG)
}
\arguments{
\item{raiD}{internal SEXP data structure}

\item{raix}{internal SEXP data structure}

\item{raing}{internal SEXP data structure}

\item{radalpha}{internal SEXP data structure}

\item{rn}{internal SEXP data structure}

\item{rp}{internal SEXP data structure}

\item{rG}{internal SEXP data structure}

\item{rS}{internal SEXP data structure}

\item{rtotiter}{internal SEXP data structure}

\item{radZ}{internal SEXP data structure}

\item{rgamma_z}{internal SEXP data structure}

\item{reta}{internal SEXP data structure}

\item{rgamma_alpha}{internal SEXP data structure}

\item{rsigma2}{internal SEXP data structure}

\item{rlambda}{internal SEXP data structure}

\item{rhyper_eta_a}{internal SEXP data structure}

\item{rhyper_eta_b}{internal SEXP data structure}

\item{rhyper_sigma2_a}{internal SEXP data structure}

\item{rhyper_sigma2_b}{internal SEXP data structure}

\item{rhyper_lambda}{internal SEXP data structure}

\item{rfamily}{internal SEXP data structure}

\item{rverbose}{internal SEXP data structure}

\item{radtheta}{internal SEXP data structure}

\item{radz}{internal SEXP data structure}

\item{radeta}{internal SEXP data structure}

\item{radsigma2}{internal SEXP data structure}

\item{radlambda}{internal SEXP data structure}

\item{radprob}{internal SEXP data structure}

\item{raix_ind}{internal SEXP data structure}

\item{rinit}{internal SEXP data structure}

\item{rM}{internal SEXP data structure}

\item{rR}{internal SEXP data structure}
}
\description{
For internal use only.
}
