% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\docType{class}
\name{dmbc_model-class}
\alias{dmbc_model-class}
\alias{dmbc_model}
\title{An S4 class to represent a DMBC model.}
\description{
An S4 class to represent a DMBC model.
}
\section{Slots}{

\describe{
\item{\code{p}}{A length-one character vector representing the number of dimensions
of the latent space to use in the MDS analysis.}

\item{\code{G}}{A length-one numeric vector representing the number of clusters to
partition the subjects into.}

\item{\code{family}}{A length-one character vector representing the type of data to
analyze. Currently, it accepts only the 'binomial' value, but future
developments will include the possibility to analyze continuous,
multinomial and count data.}
}}

\examples{
showClass("dmbc_model")

}
\references{
Venturini, S., Piccarreta, R. (2019), "A Bayesian Approach for Model-Based
  Clustering of Several Binary Dissimilarity Matrices: the \pkg{dmbc}
  Package in \code{R}", Technical report.
}
