% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_predict.r
\name{dmbc_predict}
\alias{dmbc_predict}
\title{Predict probability from DMBC model}
\usage{
dmbc_predict(data = data, testSet = testSet, auc_out = auc_out,
  col_start = 3, type_col = 2, Prior1 = 0.5, Prior2 = 1 - Prior1)
}
\arguments{
\item{data}{Validation dataset with rows are samples, columns are features. The first column should be the sample ID, second column group variable (Disease type, the label you want to classify on).}

\item{testSet}{Lable unknown testSet without sample IDs.}

\item{auc_out}{Output object of Cal_AUC() from a validation set.}

\item{col_start}{An index indicating at which column is the beginning of bacteria (features) data in the validation set. The default is the 3rd column.}

\item{type_col}{An index indicating at which column is group/type variable in the validation set. The default is the 2nd column.}

\item{Prior1}{Prevalence of label1 according to literature or experience. Default is 0.5.}

\item{Prior2}{Prevalence of label2 according to literature or experience. Default is 0.5.}
}
\value{

}
\description{
Based on a set of optimized features from training set, this function predicts the posterior probability for two class labels.
}
\examples{
#load the DMBC library
library(DMBC)

#load training dataset
data(training)

#load test dataset
data(test)

## calculate AUC based on training set using 10-fold cv ##
auc_out <- Cal_AUC(tfcv(training))

## calculate AUC based on training set using leave-one-out cv ##
auc_out <- Cal_AUC(loocv(training))

#predict unknown test set using training set and auc results.
dmbc_predict(data=training,testSet=test,auc_out=auc_out)
}
