% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.r
\name{FS}
\alias{FS}
\title{Filter and Feature Selection based on Wilcoxon Rank-Sum test}
\usage{
FS(training = data, type_col = 2, col_start = 3,
  Cutoff_mean = 0.001, Cutoff_ratio = 0.1, totalReadsCutoff = 500,
  Cutoff_pvalue = 1)
}
\arguments{
\item{training}{A data frame of training set.}

\item{type_col}{An index indicating at which column is group/type variable. The default is the 3rd column.}

\item{col_start}{An index indicating at which column is the beginning of bacteria (features) data. Default is the 2nd column.}

\item{Cutoff_mean}{The minimum average relative abundance allowed in filtering step. Default is 0.0005.}

\item{Cutoff_ratio}{The non-zero ratio cutoff in filtering features.Default value is 0.1.}

\item{totalReadsCutoff}{The minimum allowed total reads per sample. Any sample has less than this number of total reads will be removed. Default is 500.}

\item{Cutoff_pvalue}{The maximum Pvalue allowed for a givien feature to be remained on the list of the selected features.}
}
\value{
A list of 2: Feature and CountData.

\code{Feature}       A list of selected features sorted by their Wilcoxon P values.

\code{CountData}     A data frame containing balanced data.
}
\description{
Given a training set, this function performs feature selection based on several thresholds: (1). Average relative abudance in each cohort class (minimum relative abundance by default is 0.25\%), (2). Total reads per sample (minimum reads per sample is 500 by default), (3). Non-zero ratio out of all samples (By default, at least 10\% of the samples should have have non zero value.)
}
\examples{
data(training)
FS(training)
}
