% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{initialize,dmbc_fit_list-method}
\alias{initialize,dmbc_fit_list-method}
\alias{dmbc_fit_list-initialize}
\title{Create an instance of the \code{dmbc_fit_list} class using new/initialize.}
\usage{
\S4method{initialize}{dmbc_fit_list}(.Object, results = list())
}
\arguments{
\item{.Object}{Prototype object from the class \code{\link{dmbc_fit_list}}.}

\item{results}{A list whose elements are the \code{dmbc_fit} objects for
each simulated chain.}
}
\description{
Create an instance of the \code{dmbc_fit_list} class using new/initialize.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
