% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.r
\name{loocv}
\alias{loocv}
\title{Leave one out cross-validation}
\usage{
loocv(data = data, type_col = 2, col_start = 3, Cutoff_mean = 5e-04,
  Cutoff_ratio = 0.1, totalReadsCutoff = 500, Cutoff_pvalue = 0.5)
}
\arguments{
\item{data}{A data frame validation set with first column the sample IDs, second column the group variable (classificatoin variable/disease type), and third column the beginning of count table.}

\item{type_col}{An index indicating at which column is group/type variable. The default is the 3rd column.}

\item{col_start}{An index indicating at which column is the beginning of bacteria (features) data. Default is the 2nd column.}

\item{Cutoff_mean}{The minimum average relative abundance allowed in filtering step. Default is 0.0005.}

\item{Cutoff_ratio}{The non-zero ratio cutoff in filtering features.Default value is 0.1.}

\item{totalReadsCutoff}{The minimum allowed total reads per sample. Any sample has less than this number of total reads will be removed. Default is 500.}

\item{Cutoff_pvalue}{The maximal P value allowed for a given feature to be remained in the list of selected features.}
}
\value{
A list of CalPrb() results. This output will be further used for Cal_AUC() in order to estimate the final model.
}
\description{
Implementation of leave-one-out cross-validataion. It takes in an entire dataset, with first column the sample IDs, second column the group variable (classificatoin variable/disease type), using "Type" as its colum names, and third column the beginning of count table.
}
\examples{
data(training)
loocv(training)
}

