% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\docType{methods}
\name{plot,dmbc_config-method}
\alias{plot,dmbc_config-method}
\title{Provide a graphical summary of a \code{dmbc_config} class instance.}
\usage{
\S4method{plot}{dmbc_config}(x, size = NULL, size_lbl = NULL,
  adjust = 0.05, label_objects = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{dmbc_config}}.}

\item{size}{A length-two numeric vector providing the optional sizes of
points and lines in the plot.}

\item{size_lbl}{A length-one numeric vector providing the size of labels.}

\item{adjust}{A length-one numeric vector providing the optional horizontal
and vertical adjustment to nudge labels by.}

\item{label_objects}{A length-one logical vector. If \code{TRUE}, labels are
added to the plot.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Provide a graphical summary of a \code{dmbc_config} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
