% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\docType{methods}
\name{plot,dmbc_ic-method}
\alias{plot,dmbc_ic-method}
\title{Provide a graphical summary of a \code{dmbc_ic} class instance.}
\usage{
\S4method{plot}{dmbc_ic}(x, size = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{dmbc_ic}}.}

\item{size}{A length-two numeric vector providing the optional sizes of
points and lines in the plot.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Provide a graphical summary of a \code{dmbc_ic} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
