% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{show,dmbc_config-method}
\alias{show,dmbc_config-method}
\alias{dmbc_config-show}
\title{Show an instance of the \code{dmbc_config} class.}
\usage{
\S4method{show}{dmbc_config}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_config}}.}
}
\description{
Show an instance of the \code{dmbc_config} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
