% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{show,dmbc_data-method}
\alias{show,dmbc_data-method}
\alias{dmbc_data-show}
\title{Show an instance of the \code{dmbc_data} class.}
\usage{
\S4method{show}{dmbc_data}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_data}}.}
}
\description{
Show an instance of the \code{dmbc_data} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
