% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{show,dmbc_fit_list-method}
\alias{show,dmbc_fit_list-method}
\alias{dmbc_fit_list-show}
\title{Show an instance of the \code{dmbc_fit_list} class.}
\usage{
\S4method{show}{dmbc_fit_list}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_fit_list}}.}
}
\description{
Show an instance of the \code{dmbc_fit_list} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
