% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{show,dmbc_ic-method}
\alias{show,dmbc_ic-method}
\alias{dmbc_ic-show}
\title{Show an instance of the \code{dmbc_ic} class.}
\usage{
\S4method{show}{dmbc_ic}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_ic}}.}
}
\description{
Show an instance of the \code{dmbc_ic} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
