% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{show,dmbc_model-method}
\alias{show,dmbc_model-method}
\alias{dmbc_model-show}
\title{Show an instance of the \code{dmbc_model} class.}
\usage{
\S4method{show}{dmbc_model}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_model}}.}
}
\description{
Show an instance of the \code{dmbc_model} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
