% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{subset,dmbc_fit-method}
\alias{subset,dmbc_fit-method}
\alias{dmbc_fit-subset}
\title{Subsetting a \code{dmbc_fit} object.}
\usage{
\S4method{subset}{dmbc_fit}(x, pars = character(), regex_pars = character(), ...)
}
\arguments{
\item{x}{An object of class \code{\link{dmbc_fit}}.}

\item{pars}{An optional character vector of parameter names. If neither 
\code{pars} nor \code{regex_pars} is specified, the default is to use all
parameters.}

\item{regex_pars}{An optional \code{\link[=grep]{regular expression}} to use
for parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Subsetting a \code{dmbc_fit} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
