% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{summary,dmbc_config-method}
\alias{summary,dmbc_config-method}
\alias{dmbc_config-summary}
\title{Provide a summary of a \code{dmbc_config} class instance.}
\usage{
\S4method{summary}{dmbc_config}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_config}}.}
}
\description{
Provide a summary of a \code{dmbc_config} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
