% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{summary,dmbc_data-method}
\alias{summary,dmbc_data-method}
\alias{dmbc_data-summary}
\title{Provide a summary of a \code{dmbc_data} class instance.}
\usage{
\S4method{summary}{dmbc_data}(object)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_data}}.}
}
\description{
Provide a summary of a \code{dmbc_data} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
