% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{summary,dmbc_fit-method}
\alias{summary,dmbc_fit-method}
\alias{dmbc_fit-summary}
\title{Provide a summary of a \code{dmbc_fit} class instance.}
\usage{
\S4method{summary}{dmbc_fit}(object, include.burnin = FALSE, summary.Z = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_fit}}.}

\item{include.burnin}{A length-one logical vector. If \code{TRUE} the
burnin iterations (if available) are included in the summary.}

\item{summary.Z}{A length-one logical vector. If \code{TRUE} the summary
also includes the latent configuration coordinates.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Provide a summary of a \code{dmbc_fit} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
