% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{summary,dmbc_ic-method}
\alias{summary,dmbc_ic-method}
\alias{dmbc_ic-summary}
\title{Provide a summary of a \code{dmbc_ic} class instance.}
\usage{
\S4method{summary}{dmbc_ic}(object, p = NULL, G = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{dmbc_ic}}.}

\item{p}{An optional length-one numeric vector providing the number of
latent space dimension to use in the summary.}

\item{G}{An optional length-one numeric vector providing the number of
clusters to use in the summary.}
}
\description{
Provide a summary of a \code{dmbc_ic} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
