% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\docType{data}
\name{test}
\alias{test}
\title{Example Testing Data}
\format{An object of class \code{data.frame} with 6 rows and 7 columns.}
\usage{
data(test)
}
\description{
This is a simulated testing dataset with no label for each sample.
}
\examples{
data(test)
}
\keyword{datasets}

