% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.r
\docType{data}
\name{training}
\alias{training}
\title{Example Training Data}
\format{An object of class \code{data.frame} with 26 rows and 13 columns.}
\usage{
data(training)
}
\description{
This is a simulated testing dataset with label for each sample.Relative abundance is used in DMBC version 1.1.0
data(training)
data(test)
CV=loocv(training)
auc_out =Cal_AUC(CV)
dmbc_predict(data=Meta_data,testset=test,auc_out=auc_out) ##delete
dmbc_predict(data=training,testSet=test,auc_out=auc_out)
}
\examples{
data(training)
}
\keyword{datasets}
