% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbc_classes.R
\name{update}
\alias{update}
\title{Provide an update generic method for the dmbc package.}
\usage{
update(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{dmbc_ic}}.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Provide an update generic method for the dmbc package.
}
