% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\docType{data}
\name{test}
\alias{test}
\title{Example Training Data}
\format{An object of class \code{data.frame} with 5 rows and 11 columns.}
\usage{
data(test)
}
\description{
This is a simulated testing dataset with label for each sample.Relative abundance is used in DMBC version 1.1.0
}
\examples{
data(test)
}
\keyword{datasets}
