% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.r
\docType{data}
\name{training}
\alias{training}
\title{Example Training Data}
\format{An object of class \code{data.frame} with 20 rows and 9 columns.}
\usage{
data(training)
}
\description{
This is a simulated testing dataset with label for each sample.
}
\examples{
data(training)
}
\keyword{datasets}

